GHA: Change package prefix
Change the prefix in the pkg-config file such that if the Linux and
headers-only packages are extracted to `/opt` it will work out of the
box.

The Windows package has the prefix located in `/` (root of current
drive). The expectation for Windows is that the pkg-config option
`--define-prefix` shall be enabled to use the package.